docs: add story for MachineActionMenu
