+++ src/app/machines/views/MachineList/MachineListControls/MachineActionMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MachineActionMenu";
